🐛 prevent java.lang.NumberFormatException when value is "0.0"

+++ drivers/ThinQ_Washer.groovy
@@ -51,7 +51,7 @@ def processStateData(data) {
             runTime += (data["Initial_Time_H"] * 60 * 60)
             updateDataValue("initialHours", data["Initial_Time_H"].toString())
         } else {
-            runTime += (getDataValue("initialHours") ?: "0").toInteger()*60*60
+            runTime += (getDataValue("initialHours") ?: "0").toFloat().toInteger()*60*60
         }
         if (parent.checkValue(data, 'Initial_Time_M')) {
             runTime += (data["Initial_Time_M"] * 60)
@@ -67,7 +67,7 @@ def processStateData(data) {
             remainingTime += (data["Remain_Time_H"] * 60 * 60)
             updateDataValue("remainHours", data["Remain_Time_H"].toString())
         } else {
-            remainingTime += (getDataValue("remainHours") ?: "0").toInteger()*60*60
+            remainingTime += (getDataValue("remainHours") ?: "0").toFloat().toInteger()*60*60
         }
         if (parent.checkValue(data, 'Remain_Time_M')) {
             remainingTime += (data["Remain_Time_M"] * 60)
@@ -90,7 +90,7 @@ def processStateData(data) {
             delayTime += (data["Reserve_Time_H"] * 60 * 60)
             updateDataValue("reserveHours", data["Reserve_Time_H"].toString())
         } else {
-            delayTime += (getDataValue("reserveHours") ?: "0").toInteger()*60*60
+            delayTime += (getDataValue("reserveHours") ?: "0").toFloat().toInteger()*60*60
         }
         if (parent.checkValue(data, 'Reserve_Time_M')) {
             delayTime += (data["Reserve_Time_M"] * 60)
